Реализовал отправку записей на сервер и установку уникального кода устройства

+++ app/src/main/java/com/example/studassistant/entities/Group.java
@@ -1,21 +1,21 @@
 package com.example.studassistant.entities;
 
 public class Group {
-    private int id;
+    private long id;
     private String name;
 
     public Group(){}
 
-    public Group(int id, String name){
+    public Group(long id, String name){
         this.id = id;
         this.name = name;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
